+++ default.nix
@@ -7,7 +7,7 @@
 }:
 buildGoModule {
   pname = "beads";
-  version = "1.1.0";
+  version = "1.2.0";
 
   src = self;